% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combine}
\alias{combine}
\title{NFL combine data}
\format{A data frame with 4947 rows and 26 variables.}
\usage{
combine
}
\description{
Need additional description of this data
}
\keyword{datasets}
