% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAVF_breakout.R
\name{MAVF_breakout}
\alias{MAVF_breakout}
\title{Description: This function will take a matrix of SAVF scores and show the break out of each alternatives weighted SAVF scores who’s sum is their MAVF score.
Required Inputs: Matrix of SAVF scores (SAVF_matrix), numeric vector of SAVF weights, and vector or attributes names (names).
Output: MAVF breakout graph}
\usage{
MAVF_breakout(SAVF_matrix, weights, names)
}
\description{
Description: This function will take a matrix of SAVF scores and show the break out of each alternatives weighted SAVF scores who’s sum is their MAVF score.
Required Inputs: Matrix of SAVF scores (SAVF_matrix), numeric vector of SAVF weights, and vector or attributes names (names).
Output: MAVF breakout graph
}
