% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAVF_scores.R
\name{MAVF_Scores}
\alias{MAVF_Scores}
\title{Description: This function will take a matrix of SAVF scores and multiply it by the weights matrix to calculate the alternatives MAVF score.
Required Inputs: Matrix of SAVF scores (SAVF_matrix), numeric vector of SAVF weights, and vector or attributes names (names).
Output: MAVF Scores.}
\usage{
MAVF_Scores(SAVF_matrix, weights, names)
}
\description{
Description: This function will take a matrix of SAVF scores and multiply it by the weights matrix to calculate the alternatives MAVF score.
Required Inputs: Matrix of SAVF scores (SAVF_matrix), numeric vector of SAVF weights, and vector or attributes names (names).
Output: MAVF Scores.
}
