% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAVF_categorical_plot.R
\name{SAVF_cat_plot}
\alias{SAVF_cat_plot}
\title{Description: This function will plot the categorical SAVF graph. Categories may be any value, but category scores must be numeric. The function checks to ensure the total of scores sums to one.}
\usage{
SAVF_cat_plot(categories, scores)
}
\description{
Required Inputs: Vector of categories (categories) and numeric vector of category scores (scores).
}
\details{
Output: Categorical SAVF graph.
}
