% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAVF_categorical_score.R
\name{SAVF_categorical_score}
\alias{SAVF_categorical_score}
\title{Description: This function will calculate the SAVF score for a categorical value.
Required Inputs: Attribute raw value or vector or values (x), vector of categories (categories) and numeric vector of category weights (weights).
Output: Categorical SAVF Score.}
\usage{
SAVF_categorical_score(x, categories, weights)
}
\description{
Description: This function will calculate the SAVF score for a categorical value.
Required Inputs: Attribute raw value or vector or values (x), vector of categories (categories) and numeric vector of category weights (weights).
Output: Categorical SAVF Score.
}
