% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAVF_exp_plot.R
\name{SAVF_exp_plot}
\alias{SAVF_exp_plot}
\title{Description: This function will plot an increasing or decreasing exponential SAVF curve. It calls the SAVF_calc_rho and SAVF_exp_score functions and plots your score on the curve with a blue dot.}
\usage{
SAVF_exp_plot(x, x_low, x_mid, x_high, increment = 1)
}
\description{
Required Inputs: Attribute raw value (x), Lowest value (x_low), midpoint value (x_mid), highest value (x_high), Increment(1 = increasing, 2 = decreasing).
}
\details{
Output: Exponential SAVF curve with attribute plotted.
}
