% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAVF_exp_score.R
\name{SAVF_exp_score}
\alias{SAVF_exp_score}
\title{Description: This function will calculate the SAVF score for an exponentially increasing or decreasing function. It calls the SAVF_calc_rho function, so knowing rho before hand is not necessary.
Required Inputs: Attribute raw value (x), Lowest value (x_low), midpoint value (x_mid), highest value (x_high), Increment(1 = increasing, 2 = decreasing).
Output: Exponential SAVF Score}
\usage{
SAVF_exp_score(x, x_low, x_mid, x_high, increment = 1)
}
\description{
Description: This function will calculate the SAVF score for an exponentially increasing or decreasing function. It calls the SAVF_calc_rho function, so knowing rho before hand is not necessary.
Required Inputs: Attribute raw value (x), Lowest value (x_low), midpoint value (x_mid), highest value (x_high), Increment(1 = increasing, 2 = decreasing).
Output: Exponential SAVF Score
}
