% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAVF_liner_plot.R
\name{SAVF_lin_plot}
\alias{SAVF_lin_plot}
\title{Description: This function will calculate the linear SAVF curve for an increasing or decreasing function. It calls the SAVF_linear_score function to calculate the score based on the midpoint value method and plots it with a blue dot.}
\usage{
SAVF_lin_plot(x, x_low, x_mid, x_high, increment = 1)
}
\description{
Required Inputs: Attribute raw value (x), Lowest value (x_low), midpoint value (x_mid), highest value (x_high), Increment(1 = increasing, 2 = decreasing).
}
\details{
Output: Linear SAVF curve with attribute plotted.
}
