% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAVF_linear_plot.R
\name{SAVF_linear_plot}
\alias{SAVF_linear_plot}
\title{Single Attribute Value Function (SAVF) Linear Plot}
\usage{
SAVF_linear_plot(x, x_low, x_mid, x_high, increasing = TRUE)
}
\arguments{
\item{x}{Attribute raw value}

\item{x_low}{Lowest value}

\item{x_mid}{Midpoint value}

\item{x_high}{Highest value}

\item{increasing}{TRUE=increasing, FALSE=decreasing, Default: TRUE}
}
\value{
Linear SAVF curve with attribute plotted
}
\description{
: Plots the linear Single Attribute Value Function (SAVF) graph for an increasing or decreasing function. 
It calls the SAVF_linear_score function to calculate the score based on the midpoint value method and plots it with a blue dot.
}
\examples{
\dontrun{ SAVF_linear_plot(10, 0, 25, 100, FALSE)}

}
