% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_plot.R
\name{sensitivity_plot}
\alias{sensitivity_plot}
\title{Description: This function will take a matrix of SAVF scores and show how each alternatives MAVF scores change as the weight for that criteria changes from zero to one. The vertical black line represents the current weight.
Required Inputs: Matrix of SAVF scores (SAVF_matrix), numeric vector of SAVF weights, the alternativenames (names), and which criteria to conduct sensitivity analysis on (criteria).
Output: Sensitivity Analysis graph}
\usage{
sensitivity_plot(SAVF_matrix, weights, names, criteria)
}
\description{
Description: This function will take a matrix of SAVF scores and show how each alternatives MAVF scores change as the weight for that criteria changes from zero to one. The vertical black line represents the current weight.
Required Inputs: Matrix of SAVF scores (SAVF_matrix), numeric vector of SAVF weights, the alternativenames (names), and which criteria to conduct sensitivity analysis on (criteria).
Output: Sensitivity Analysis graph
}
