% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAFV_calc_rho.R
\name{SAVF_calc_rho}
\alias{SAVF_calc_rho}
\title{Description: The function above calculates rho for an exponential function.
Required inputs:Lowest Value (x_low), Midpoint Value (x_mid), Highest Value (x_high), Increment(1 = increasing, 2 = decreasing).
Output: Rho}
\usage{
SAVF_calc_rho(x_low, x_mid, x_high, increment = 1)
}
\description{
Description: The function above calculates rho for an exponential function.
Required inputs:Lowest Value (x_low), Midpoint Value (x_mid), Highest Value (x_high), Increment(1 = increasing, 2 = decreasing).
Output: Rho
}
