% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAVF_linear_score.R
\name{SAVF_linear_score}
\alias{SAVF_linear_score}
\title{Description: This function will calculate the SAVF score for a linearly increasing or decreasing function. It calculates the score based on the midpoint value method.
Required Inputs: Attribute raw value (x), Lowest value (x_low), midpoint value (x_mid), highest value (x_high), Increment(1 = increasing, 2 = decreasing).
Output: Linear SAVF Score.}
\usage{
SAVF_linear_score(x, x_low, x_mid, x_high, increment = 1)
}
\description{
Description: This function will calculate the SAVF score for a linearly increasing or decreasing function. It calculates the score based on the midpoint value method.
Required Inputs: Attribute raw value (x), Lowest value (x_low), midpoint value (x_mid), highest value (x_high), Increment(1 = increasing, 2 = decreasing).
Output: Linear SAVF Score.
}
